requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-docx==1.1.0
python-multipart==0.0.6
docx2pdf==0.1.8
pypandoc==1.12
Pillow==10.2.0
pdf2image==1.17.0
aiofiles==23.2.1
